requirements.txt
flask
python-dotenv
requests
# cryptography (opcional se usar certificados/autos)
